+++ dask_match/io/__init__.py
@@ -1,1 +1,3 @@
+from dask_match.io.csv import *
+from dask_match.io.io import *
 from dask_match.io.parquet import *
