+++ dask_match/io/io.py
@@ -1,0 +1,65 @@
+import math
+
+from dask_match.core import Expr, new_collection
+
+
+class IO(Expr):
+    pass
+
+class FromPandas(IO):
+    """The only way today to get a real dataframe"""
+
+    _parameters = ["frame", "npartitions"]
+    _defaults = {"npartitions": 1}
+
+    @property
+    def _meta(self):
+        return self.frame.head(0)
+
+    def _divisions(self):
+        return [None] * (self.npartitions + 1)
+
+    def _layer(self):
+        chunksize = int(math.ceil(len(self.frame) / self.npartitions))
+        locations = list(range(0, len(self.frame), chunksize)) + [len(self.frame)]
+        return {
+            (self._name, i): self.frame.iloc[start:stop]
+            for i, (start, stop) in enumerate(zip(locations[:-1], locations[1:]))
+        }
+
+    def __str__(self):
+        return "df"
+
+    __repr__ = __str__
+
+
+def from_pandas(*args, **kwargs):
+    return new_collection(FromPandas(*args, **kwargs))
+
+
+class FromGraph(IO):
+    """A DataFrame created from an opaque Dask task graph
+
+    This is used in persist, for example, and would also be used in any
+    conversion from legacy dataframes.
+    """
+
+    _parameters = ["layer", "_meta", "divisions", "_name"]
+
+    @property
+    def _meta(self):
+        return self.operand("_meta")
+
+    def _divisions(self):
+        return self.operand("divisions")
+
+    @property
+    def _name(self):
+        return self.operand("_name")
+
+    def _layer(self):
+        return self.operand("layer")
+
+
+def from_graph(*args, **kwargs):
+    return new_collection(FromGraph(*args, **kwargs))
