+++ dask_match/io/parquet.py
@@ -14,7 +14,8 @@ from dask.utils import natural_sort_key
 from fsspec.utils import stringify_path
 from matchpy import CustomConstraint, Pattern, ReplacementRule, Wildcard
 
-from dask_match.core import EQ, GE, GT, IO, LE, LT, NE, Filter, new_collection
+from dask_match.core import EQ, GE, GT, LE, LT, NE, Filter, new_collection
+from dask_match.io import IO
 
 NONE_LABEL = "__null_dask_index__"
